CTM: begin solver rewrite
